refactor linkedinScraper in puppeteer service search and scrap than return an array ofscraped profils data
