+++ src/models/profile.model.ts
@@ -5,7 +5,7 @@ import mongoose from 'mongoose';
 //   profileName: string;
 //   position: string;
 //   email: string;
-//   linkdinLink: string;
+//   profileLink: string;
 //   isEmailSent: boolean;
 //   isStared: boolean;
 //   isVmarked: boolean;
@@ -18,16 +18,15 @@ import mongoose from 'mongoose';
 export interface ProfileDocument extends mongoose.Document {
   //userId: typeof mongoose.Types.ObjectId;
   name: string;
-  profileName: string;
   position: string;
   email: string;
-  linkdinLink: string;
-  isEmailSent: boolean;
-  isStared: boolean;
-  isVmarked: boolean;
+  profileLink: string;
   tags: string[];
   imageSrc: string;
   comment: string;
+  isEmailSent: boolean;
+  isStared: boolean;
+  isVmarked: boolean;
   createdAt: Date
   updatedAt: Date
 }
@@ -38,16 +37,27 @@ const profileSchema:mongoose.Schema<ProfileDocument> = new mongoose.Schema({
   //   // required
   // },
   name: String,
-  profileName: String,
   position: String,
   email: String,
-  linkdinLink: String,
-  isEmailSent: Boolean,
-  isStared: Boolean,
-  isVmarked: Boolean,
+  profileLink: {
+    type: String,
+    unique: true,
+  },
   tags: [String],
   imageSrc: String,
   comment: String,
+  isEmailSent: {
+    type: Boolean,
+    default: false,
+  },
+  isStared: {
+    type: Boolean,
+    default: false,
+  },
+  isVmarked: {
+    type: Boolean,
+    default: false,
+  },
 }, {
   timestamps: true,
 });
