+++ src/services/puppeteer.service.ts
@@ -2,6 +2,7 @@ import puppeteer, { Page } from 'puppeteer';
 import { config } from 'dotenv';
 import * as path from 'path';
 import * as utils from '../utils/puppeteer.util';
+import { ProfileDocument } from '../models/profile.model';
 
 config({ path: path.resolve(__dirname, '../.env') });
 
@@ -12,18 +13,25 @@ export const linkedinScraper = async () => {
     browser = await puppeteer.launch({
       headless: false,
       defaultViewport: null,
-      slowMo: utils.getRandomTimeInterval(500),
+      slowMo: utils.getRandomTimeInterval(750, 750),
     });
 
     const page: Page = await browser.newPage();
     await utils.setLinkedinCookies(page);
 
-    const result = await getSearchResults(page);
-    console.log("result.length = ", result.length);
+    const scrapedProfiles: ProfileDocument[] = [];
+    const profilesURLs = await getSearchResults(page);
+
+    for (const profile of profilesURLs) {
+      const profileData = await scrapeSingleProfile(page, profile);
+      if (profileData) {
+        scrapedProfiles.push(profileData);
+      }
+    }
     
-    console.log(result);
+    console.log(scrapedProfiles);
     
-    return result;
+    return scrapedProfiles;
   } catch (error) {
     console.error(error);
     return [];
@@ -41,7 +49,7 @@ async function getSearchResults(page: Page): Promise<any> {
     let isNext: boolean = true;
 
     while (isNext) {
-      await page.waitForTimeout(Math.random() * 30000 + 10000);
+      await page.waitForTimeout(utils.getRandomTimeInterval(30000, 10000));
       await utils.scrollToBottom(page);
       
       const pageResult: string[] = await scrapeProfilesLinks(page);
@@ -49,6 +57,8 @@ async function getSearchResults(page: Page): Promise<any> {
   
       const nextPageButton = await getNextButton(page);
       if (!nextPageButton) {
+        await utils.getLinkedinCookies(page);
+
         return searchResults.filter((result) => {
           return result.includes('miniProfile');
         });
@@ -98,6 +108,51 @@ async function getNextButton(page: Page) {
   }
 
   return nextPageButton;
+}
+
+
+async function scrapeSingleProfile(page: Page, url: string): Promise<ProfileDocument | null> {
+  
+  await page.goto(url);
+  await page.waitForTimeout(utils.getRandomTimeInterval(30000, 10000));
+
+  const randTime = utils.getRandomTimeInterval(1000, 2000);
+  try {
+    const res = await page.evaluate(async (randTime) => {
+      const profileData = <ProfileDocument>{};
+  
+      const imageSrc = document.querySelector('div > div > div > button > img');
+      if (imageSrc) profileData.imageSrc = (imageSrc as HTMLImageElement).src;
+    
+      const name =  document.querySelector('h1'); 
+      if (name) profileData.name = name.innerText;
+    
+      const position = document.querySelector('div.text-body-medium.break-words');
+      if (position) profileData.position = (position as HTMLElement).innerText; 
+      
+      const infoButton = document.querySelector('#top-card-text-details-contact-info');
+      if (infoButton) {
+        (infoButton as HTMLButtonElement).click();
+        await new Promise(function(resolve) {setTimeout(resolve, randTime)});
+
+        const profileLink = document.querySelector('section.ci-vanity-url > div > a'); 
+        if (profileLink) profileData.profileLink = (profileLink as HTMLAnchorElement).href;
+
+        const email = document.querySelector('section.ci-email > div > a');
+        if (email) profileData.email = (email as HTMLAnchorElement).href;
+      }
+      
+      const closeModalButton = document.querySelector('div[role=dialog] > button');
+      if (closeModalButton) (closeModalButton as HTMLButtonElement).click();
+
+      return profileData;
+    }, randTime);
+  
+    return res;
+  } catch (error) {
+    console.error(error);
+    return null;
+  }
 }
